[CHANGE] Sumoselect JS updated to the latest version
