+++ aa_bulletin_board/templates/aa_bulletin_board/bundles/sumoselect-css.html
@@ -1,3 +1,3 @@
 {% load static %}
 
-<link rel="stylesheet" href="{% static 'aa_bulletin_board/libs/sumoselect/3.4.8/sumoselect.min.css' %}">
+<link rel="stylesheet" href="{% static 'aa_bulletin_board/libs/sumoselect/3.4.9/sumoselect.min.css' %}">
