+++ aa_bulletin_board/templates/aa_bulletin_board/bundles/sumoselect-js.html
@@ -1,3 +1,3 @@
 {% load static %}
 
-<script src="{% static 'aa_bulletin_board/libs/sumoselect/3.4.8/jquery.sumoselect.min.js' %}"></script>
+<script src="{% static 'aa_bulletin_board/libs/sumoselect/3.4.9/jquery.sumoselect.min.js' %}"></script>
